+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/listeners/DefaultCompetitionChangeListener.java
@@ -20,7 +20,10 @@ abstract public class DefaultCompetitionChangeListener implements CompetitionCha
     for (Integer competedGameId : competedGameIds) {
       Game game = gameService.getGame(competedGameId);
       if (game != null) {
-        frontendStatusService.deAugmentWheel(game);
+        boolean stillActive = competitionService.getCompetitionForGame(competedGameId).stream().anyMatch(Competition::isActive);
+        if (!stillActive) {
+          frontendStatusService.deAugmentWheel(game);
+        }
       }
     }
   }